meta-qti-security: Remove GCC version hard-coding from dlkm recipe
Removed hard-coding of GCC version from securemsmdlkm recipe.
We use the python filter function, which checks each element of
string variable GCCVERSION to strip the '%' symbol from it. The final
value is stored in GCCVER_AVAILABLE appended with .0, e.g. 11.3.0.
This change is required to allow for seamless Poky version upgrades.
Reference change: 4703690

Test: Validated securemsm drivers present on Yocto Kirkstone LE targets.

Change-Id: If3317234ef4eb669d9dc770ec6ed65e35119b8d4