meta-qti-security: Remove GCC version hard-coding from dlkm recipe
Removed hard-coding of GCC version from securemsmdlkm recipe.
We use the python filter function, which checks each element of
string variable GCCVERSION to strip the '%' symbol from it. The final
value is stored in GCCVER_AVAILABLE appended with .0, e.g. 11.3.0.
This change is required to allow for seamless Poky version upgrades.
Reference change: 4703690

Test: Validated securemsm drivers present on LE target.

Change-Id: I7043380fcd1c20f467cae720b577b259e5898ce9