Adding build variant in recipes to build kernel modules outside kernel.
In bazel build_module.sh , variable VARIANT is always defaults to
debug-defconfig , irespective of perf/debug variant is chosen for
vendor build, because of this "debug Module.symvers" is generated
for this module build and debug CRC of symbol module_layout is
captured in .ko.

Now, if we have perf build flashed on the device and
CONFIG_MODVERSIONS is enabled,insmoding of external module
will cause CRC mismatch for symbol module_layout causing
insmod failure.
if we have debug buid flashed CRC check for module_layout will
be success.

As a fix, (as we have prebuild kernel) we will pass build VARIENT
information from vendor build to build_modle.sh from dlkm recipes.

Change-Id: If1432444584b6617a05f7c57e205c7047a9ba7a1